# src/main/java/org/example/carrentalsystem/dto/payment/PaymentCreateRequest.java
package org.example.carrentalsystem.dto.payment;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import org.example.carrentalsystem.enums.PaymentMethod;


@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class PaymentCreateRequest {

    @NotNull
    private PaymentMethod paymentMethod;

}