# src/main/resources/api-docs.yaml
openapi: 3.1.0
info:
  title: Car Rental API
  description: REST API for car rental service
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/v1/payments/booking/{bookingId}:
    get:
      tags:
      - Payments
      summary: Get payments by booking ID
      description: Returns all payments for the specified booking. Available to the
        booking owner
      operationId: getByBookingId
      parameters:
      - name: bookingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Payments successfully retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentResponse"
        "404":
          description: Booking not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Payments
      summary: Create payment for booking
      description: Creates a new payment for the specified booking. Available to the
        booking owner
      operationId: create_2
      parameters:
      - name: bookingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCreateRequest"
        required: true
      responses:
        "201":
          description: Payment successfully created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          description: Invalid payment data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "404":
          description: Booking not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
...
  /api/v1/bookings:
    get:
      tags:
      - Bookings
      summary: Get all bookings
      description: Returns all bookings. Available only to administrators
      operationId: getAll_3
      responses:
        "200":
          description: Bookings successfully retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Bookings
      summary: Create booking
      description: Creates a new booking for the authenticated user
      operationId: create_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingCreateRequest"
        required: true
      responses:
        "201":
          description: Booking successfully created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          description: Invalid booking data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "409":
          description: Car is already booked for the selected period
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
      security:
      - bearerAuth: []
...
  /api/v1/bookings/{id}/cancel:
    patch:
      tags:
      - Bookings
      summary: Cancel booking
      description: Cancels a booking belonging to the authenticated user
      operationId: cancel
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Booking successfully cancelled
        "403":
          description: User does not have access to this booking
        "404":
          description: Booking not found
        "409":
          description: Booking cannot be cancelled
      security:
      - bearerAuth: []
...
  /api/v1/bookings/{id}:
    get:
      tags:
      - Bookings
      summary: Get bookings by id
      description: Returns booking by id. Available only to administrators
      operationId: getById_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Booking successfully retrieved
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingResponse"
      security:
      - bearerAuth: []
  /api/v1/bookings/user/{userId}:
    get:
      tags:
      - Bookings
      summary: Get bookings by user id
      description: Returns all user bookings. Available only to administrators
      operationId: getByUserId
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Bookings successfully retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
        "403":
          description: Access denied
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
      security:
      - bearerAuth: []
  /api/v1/bookings/my:
    get:
      tags:
      - Bookings
      summary: Get my bookings
      description: Returns bookings of the authenticated user
      operationId: getMyBookings
      responses:
        "200":
          description: Bookings successfully retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingResponse"
      security:
      - bearerAuth: []
components:
  schemas:
...
    PaymentCreateRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum:
          - CARD
          - CASH
      required:
      - paymentMethod
    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookingId:
          type: integer
          format: int64
        amount:
          type: number
        status:
          type: string
          enum:
          - PENDING
          - COMPLETED
          - FAILED
          - REFUNDED
        paymentMethod:
          type: string
          enum:
          - CARD
          - CASH
        transactionId:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    BookingCreateRequest:
      type: object
      properties:
        carId:
          type: integer
          format: int64
        pickupLocationId:
          type: integer
          format: int64
        returnLocationId:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
      - carId
      - endDate
      - pickupLocationId
      - returnLocationId
      - startDate
    BookingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        carId:
          type: integer
          format: int64
        carBrand:
          type: string
        carModel:
          type: string
        pickupLocationId:
          type: integer
          format: int64
        pickupLocationName:
          type: string
        returnLocationId:
          type: integer
          format: int64
        returnLocationName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalPrice:
          type: number
        status:
          type: string
          enum:
          - PENDING
          - CONFIRMED
          - CANCELLED
          - COMPLETED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT